+++ src/Application/Components/Import/ParticipantImportTag.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Components\Import;
 
 final class ParticipantImportTag
@@ -19,4 +11,5 @@ final class ParticipantImportTag
     public const PARTICIPANT_IMPORT_FILE = 'participant_import_file';
     public const PARTICIPANT_IMPORT_CHARSET = 'participant_import_charset';
     public const PARTICIPANT_IMPORT_ALLOW_MULTI_SHEET = 'participant_import_allow_multi_sheet';
+    public const PARTICIPANT_IMPORT_SAVED_MAPPING = 'participant_import_saved_mapping';
 }